+++ RELEASE_NOTES.rst
@@ -2,6 +2,13 @@
 Release notes for parquery
 ========================
 
+Release  2.0.0
+==============
+- Remove dependencies on Numpy and pandas
+- Fully switch to pyarrow for all data handling and aggregation
+- Drop support for python 2.7 and 3.6, require minimum python
+- Add optional support for Pandas & Polars for specific functions
+
 Release  1.2.0
 ==============
 - Remove python 3.7 pipelines and add python 3.11 support